bot/requirements.txt
python-telegram-bot[job-queue]>=20.0,<21.0
anthropic>=0.25,<1.0
supabase>=2.0,<3.0
python-dotenv>=1.0,<2.0
tzdata>=2024.1
